+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KQSSTN54CNYW431XRESGQHB2"
+        "01KQVKDCKD4EPZWYAX4068J66G"
     )
     r"""The unique identifier of the data source"""