+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KQSSTN55GDVW1V74DQ93P6ST"
+        "01KQVKDCKSKTVB45F5VE2CYJW3"
     )
     r"""The unique identifier of the data source"""