+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KQSSTN582C3J248Q3D6RACVV"
+        "01KQVKDCMGA3TE6MA46HBTPE76"
     )
     r"""The unique identifier of the data source"""