+++ packages/orq-rc/src/orq_ai_sdk/models/updatetoolop.py
@@ -292,7 +292,7 @@ class RequestBodyTools(BaseModel):
 
     schema_: Annotated[UpdateToolRequestBodyToolsSchema, pydantic.Field(alias="schema")]
 
-    id: Optional[str] = "01KQSSTMT2GXXZRRPPXE50VEGJ"
+    id: Optional[str] = "01KQVKDC9RWGHVF6X7P80SM1T5"
 
     description: Optional[str] = None
 
@@ -1357,7 +1357,7 @@ class UpdateToolResponseBodyCodeExecutionTool(BaseModel):
     code_tool: UpdateToolResponseBodyCodeTool
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KQSSTMRH4EQ1ACVSJCVHE8FQ"
+        "tool_01KQVKDC9J1Z37DC1Z9T4ZDKDB"
     )
 
     display_name: Optional[str] = None
@@ -1479,7 +1479,7 @@ class UpdateToolResponseBodyTools(BaseModel):
         UpdateToolResponseBodyToolsSchema, pydantic.Field(alias="schema")
     ]
 
-    id: Optional[str] = "01KQSSTMRBRDSEW9C49GX42RZA"
+    id: Optional[str] = "01KQVKDC9JPPE27R0ZCC47JD2F"
 
     description: Optional[str] = None
 
@@ -1623,7 +1623,7 @@ class UpdateToolResponseBodyMCPTool(BaseModel):
     mcp: UpdateToolResponseBodyMcp
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KQSSTMR43C1FXDF5FEB82GH9"
+        "tool_01KQVKDC9GMNC68JJY8KXTFP3Q"
     )
 
     display_name: Optional[str] = None
@@ -1914,7 +1914,7 @@ class UpdateToolResponseBodyHTTPTool(BaseModel):
     http: UpdateToolResponseBodyHTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KQSSTMQQ3QANCY31SPC38YKA"
+        "tool_01KQVKDC9F1Q9P6J19FC4FN544"
     )
 
     display_name: Optional[str] = None
@@ -2099,7 +2099,7 @@ class UpdateToolResponseBodyJSONSchemaTool(BaseModel):
     json_schema: UpdateToolResponseBodyJSONSchema
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KQSSTMQCSRZM94G01DGTFT8J"
+        "tool_01KQVKDC9DVQBAGNZ27VE47HEG"
     )
 
     display_name: Optional[str] = None
@@ -2288,7 +2288,7 @@ class UpdateToolResponseBodyFunctionTool(BaseModel):
     function: UpdateToolResponseBodyFunction
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KQSSTMQ89BZ86B8ER966R6CE"
+        "tool_01KQVKDC9C4BB1YC0VJBGRWA78"
     )
 
     display_name: Optional[str] = None