+++ packages/orq-rc/RELEASES.md
@@ -7559,3 +7559,13 @@ Based on:
 - [python v4.9.0-rc.23] packages/orq-rc
 ### Releases
 - [PyPI v4.9.0-rc.23] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.23 - packages/orq-rc
+
+## 2026-05-05 08:20:29
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.11 (2.881.17) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.9.0-rc.24] packages/orq-rc
+### Releases
+- [PyPI v4.9.0-rc.24] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.24 - packages/orq-rc